decouple ClassMethodVisibilityVendorLockResolver
